+++ app/routes/__layout/admin/groups.tsx
@@ -1,6 +1,6 @@
 import type { LoaderArgs } from "@remix-run/node"
 import { json } from "@remix-run/node"
-import { useLoaderData } from "@remix-run/react"
+import { Link, useLoaderData } from "@remix-run/react"
 import styled from "styled-components"
 import { HoverCard } from "~/components/HoverCard"
 import { Table } from "~/components/Table"
@@ -20,7 +20,7 @@ export default function AdminGroupsPage() {
 
   return (
     <div>
-      <Title>Groups</Title>
+      <Title>Clubs</Title>
       <Table>
         <Table.Head>
           <tr>
@@ -28,13 +28,15 @@ export default function AdminGroupsPage() {
             <Table.Heading>Members</Table.Heading>
             <Table.Heading>Locations</Table.Heading>
             <Table.Heading>Lunches</Table.Heading>
-            <Table.Heading>Scores</Table.Heading>
+            <Table.Heading>Ratings</Table.Heading>
           </tr>
         </Table.Head>
         <tbody>
           {groups.map((group) => (
             <tr key={group.id}>
-              <Table.Cell title={group.id}>{group.name}</Table.Cell>
+              <Table.Cell title={group.id}>
+                <Link to={`/groups/${group.id}`}>{group.name}</Link>
+              </Table.Cell>
               <Table.Cell numeric>
                 <HoverCard>
                   <HoverCard.Trigger>{group.members.length}</HoverCard.Trigger>