+++ app/routes/__layout/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2022-09-11
+
+- **Improved landing page**
+
+  - Updated the landing page to be a bit clearer and more inviting.
+
+- **Clearer naming**
+
+  - Renamed "Score" to "Rating"
+  - Renamed "Group" to "Club"
+
 ## 2022-09-04
 
 ### New features
@@ -27,15 +38,15 @@
 
 ### Bug fixes
 
-- Redirect to lunch when submitting requested score
+- Redirect to lunch when submitting requested rating
 - Fix issue where help tooltips didn't show on mobile
 
 ## 2022-08-25
 
 ### New features
 
-- **Score requests**
-  - You can now request a score from another member in your group. They will get a notification on their start page.
+- **Rating requests**
+  - You can now request a rating from another member in your group. They will get a notification on their start page.
 
 ## 2022-08-23